refactor(main): M-5 extract smoke/QA orchestration into src/main/smoke behind DSH_SMOKE gate
